Finally a working implementation. Also with z-buffering. Change to fixedpoint later.

+++ perspective_texture_mapping/rasterizer.h
@@ -3,15 +3,15 @@
 #include <linealg.h>
 
 void DrawTriangle(
-		  std::vector<Vector4i>& vertexData,
-		  std::vector<Vector4i>& textureData,
+		  std::vector<Vector4f>& vertexData,
+		  std::vector<Vector4f>& textureData,
 		  unsigned int* buffer,
 		  unsigned int width,
 		  unsigned int height
 		  );
 void TriangleSplit(
-		   std::vector<Vector4i>& vertexData,
-		   std::vector<Vector4i>& textureData
+		   std::vector<Vector4f>& vertexData,
+		   std::vector<Vector4f>& textureData
 		   );
 #endif
